Added options and arguments
It will now support options (flags) and arguemnts. f.eks -dd -n test -e test@test.test  to add a person with the name test and email test@test.test 

It will then exit. Fastest way to add. 

It also supports commands like -g to print the entire book. It is easy to upgrade this functionality, and add more commands. But be aware of the bad implementation of the searching for the commands - since -a will be true for -add and -a